Added checks before dropping coloumns

+++ database/migrations/2021_03_30_090722_add_homebase_to_users_table.php
@@ -25,8 +25,10 @@ class AddHomebaseToUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('homebase');
-        });
+        if (Schema::hasColumn('users', 'homebase')){
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('homebase');
+            });
+        }
     }
 }
